Support three dimensional geojson
According to geojson spec the coordinates of a feature can have 3 dimensions.
Before this change the 3rd component of the geometry was misused to store the simplification values of the Douglas-Peucker algorithm.
With this change the simplification values are stored in a separate array(sqDist) and the 3rd component of the coordinate remains untouched.

Signed-off-by: Nino Kettlitz <1396039+ninok@users.noreply.github.com>